Refactor: pgerrorパッケージの場所を移動
