+++ internal/infrastructure/rdb/query_service/user/user_query_service.go
@@ -5,7 +5,7 @@ import (
 	"context"
 
 	"go-boilerplate/internal/infrastructure/rdb/driver/loggingdb"
-	"go-boilerplate/internal/infrastructure/rdb/postgres/pgerror"
+	"go-boilerplate/internal/infrastructure/rdb/pgerror"
 	"go-boilerplate/internal/infrastructure/rdb/sqlc"
 	"go-boilerplate/internal/infrastructure/rdb/sqlc/gen"
 	"go-boilerplate/internal/observability"
