+++ internal/infrastructure/rdb/system_query/healthcheck/health_check_system_query.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"go-boilerplate/internal/infrastructure/rdb/driver/loggingdb"
-	"go-boilerplate/internal/infrastructure/rdb/postgres/pgerror"
+	"go-boilerplate/internal/infrastructure/rdb/pgerror"
 	"go-boilerplate/internal/infrastructure/rdb/sqlc/gen"
 	"go-boilerplate/internal/observability"
 	"go-boilerplate/internal/usecase/healthcheck/query"
